+++ MeetingBar/AppDelegate.swift
@@ -150,6 +150,14 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
         Defaults[.hideMeetingTitle].toggle()
     }
 
+    @objc
+    func refreshSources() {
+//        TODO: fix refresh sources
+//        statusBarItem.eventStore.refreshSourcesIfNecessary()
+        statusBarItem.updateTitle()
+        statusBarItem.updateMenu()
+    }
+
     func setup() {
         statusBarItem = StatusBarItemController()
         statusBarItem.setAppDelegate(appdelegate: self)
@@ -192,24 +200,28 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
 
         showEventDetailsObserver = Defaults.observe(.showEventDetails) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change showEventDetails from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateMenu()
             }
         }
 
         shortenEventTitleObserver = Defaults.observe(.shortenEventTitle) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change shortenEventTitle from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateMenu()
             }
         }
 
         menuEventTitleLengthObserver = Defaults.observe(.menuEventTitleLength) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change menuEventTitleLengthObserver from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateMenu()
             }
         }
 
         meetingTitleVisibilityObserver = Defaults.observe(.hideMeetingTitle) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change hideMeetingTitle from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateMenu()
                 self.statusBarItem.updateTitle()
 
@@ -223,30 +235,35 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
 
         showEventEndTimeObserver = Defaults.observe(.showEventEndTime) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change showEventEndTime from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateMenu()
             }
         }
 
         eventTimeFormatObserver = Defaults.observe(.eventTimeFormat) { change in
             if change.oldValue != change.newValue {
+                NSLog("Changed eventTimeFormat from \(String(describing: change.oldValue)) to \(String(describing: change.newValue))")
                 self.statusBarItem.updateTitle()
             }
         }
 
         statusbarEventTitleLengthObserver = Defaults.observe(.statusbarEventTitleLength) { change in
             if change.oldValue != change.newValue {
+                NSLog("Changed statusbarEventTitleLengthLimits from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateTitle()
             }
         }
 
         disablePastEventObserver = Defaults.observe(.disablePastEvents) { change in
             if change.oldValue != change.newValue {
+                NSLog("Changed disablePastEvents from \(String(describing: change.oldValue)) to \(String(describing: change.newValue))")
                 self.statusBarItem.updateMenu()
             }
         }
 
         showPendingEventObserver = Defaults.observe(.showPendingEvents) { change in
             if change.oldValue != change.newValue {
+                NSLog("Changed showPendingEvents from \(String(describing: change.oldValue)) to \(String(describing: change.newValue))")
                 self.statusBarItem.updateTitle()
                 self.statusBarItem.updateMenu()
             }
@@ -254,18 +271,21 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
 
         selectedCalendarIDsObserver = Defaults.observe(.selectedCalendarIDs) { change in
             if change.oldValue != change.newValue {
+                NSLog("Changed selectedCalendarIDs from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.loadCalendars()
             }
         }
 
         showMeetingServiceIconObserver = Defaults.observe(.showMeetingServiceIcon) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change showMeetingServiceIcon from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateMenu()
             }
         }
 
         allDayEventsObserver = Defaults.observe(.allDayEvents) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change allDayEvents from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateTitle()
                 self.statusBarItem.updateMenu()
             }
@@ -273,44 +293,69 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
 
         nonAllDayEventsObserver = Defaults.observe(.nonAllDayEvents) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change nonAllDayEvents from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateTitle()
                 self.statusBarItem.updateMenu()
             }
         }
 
-        timeFormatObserver = Defaults.observe(.timeFormat) { _ in
-            self.statusBarItem.updateMenu()
+        timeFormatObserver = Defaults.observe(.timeFormat) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Change timeFormat from \(change.oldValue) to \(change.newValue)")
+                self.statusBarItem.updateMenu()
+            }
         }
 
         bookmarksObserver = Defaults.observe(keys: .bookmarks) {
             self.statusBarItem.updateMenu()
         }
 
-        eventTitleFormatObserver = Defaults.observe(.eventTitleFormat) { _ in
-            self.statusBarItem.updateTitle()
-            self.statusBarItem.updateMenu()
+        eventTitleFormatObserver = Defaults.observe(.eventTitleFormat) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Changed eventTitleFormat from \(String(describing: change.oldValue)) to \(String(describing: change.newValue))")
+                self.statusBarItem.updateTitle()
+                self.statusBarItem.updateMenu()
+            }
         }
 
-        eventTitleIconFormatObserver = Defaults.observe(.eventTitleIconFormat) { _ in
-            self.statusBarItem.updateTitle()
+        eventTitleIconFormatObserver = Defaults.observe(.eventTitleIconFormat) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Changed eventTitleFormat from \(String(describing: change.oldValue)) to \(String(describing: change.newValue))")
+                self.statusBarItem.updateTitle()
+            }
         }
 
-        pastEventsAppereanceObserver = Defaults.observe(.pastEventsAppereance) { _ in
-            self.statusBarItem.updateMenu()
+        pastEventsAppereanceObserver = Defaults.observe(.pastEventsAppereance) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Changed pastEventsAppereance from \(change.oldValue) to \(change.newValue)")
+                self.statusBarItem.updateMenu()
+            }
         }
-        declinedEventsAppereanceObserver = Defaults.observe(.declinedEventsAppereance) { _ in
-            self.statusBarItem.updateMenu()
+        declinedEventsAppereanceObserver = Defaults.observe(.declinedEventsAppereance) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Changed declinedEventsAppereance from \(change.oldValue) to \(change.newValue)")
+                self.statusBarItem.updateMenu()
+            }
         }
-        personalEventsAppereanceObserver = Defaults.observe(.personalEventsAppereance) { _ in
-            self.statusBarItem.updateTitle()
-            self.statusBarItem.updateMenu()
+        personalEventsAppereanceObserver = Defaults.observe(.personalEventsAppereance) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Changed personalEventsAppereance from \(change.oldValue) to \(change.newValue)")
+                self.statusBarItem.updateTitle()
+                self.statusBarItem.updateMenu()
+            }
         }
-        showEventsForPeriodObserver = Defaults.observe(.showEventsForPeriod) { _ in
-            self.statusBarItem.updateTitle()
-            self.statusBarItem.updateMenu()
+        showEventsForPeriodObserver = Defaults.observe(.showEventsForPeriod) { change in
+            if change.oldValue != change.newValue {
+                NSLog("Changed showEventsForPeriod from \(change.oldValue) to \(change.newValue)")
+                self.statusBarItem.updateTitle()
+                self.statusBarItem.updateMenu()
+            }
         }
         launchAtLoginObserver = Defaults.observe(.launchAtLogin) { change in
-            SMLoginItemSetEnabled(AutoLauncher.bundleIdentifier as CFString, change.newValue)
+            if change.oldValue != change.newValue {
+                NSLog("Changed launchAtLogin from \(change.oldValue) to \(change.newValue)")
+                SMLoginItemSetEnabled(AutoLauncher.bundleIdentifier as CFString, change.newValue)
+            }
         }
         preferredLanguageObserver = Defaults.observe(.preferredLanguage) { change in
             if I18N.instance.changeLanguage(to: change.newValue) {
@@ -319,21 +364,26 @@ class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDele
             }
         }
         joinEventNotificationObserver = Defaults.observe(.joinEventNotification) { change in
-            if change.newValue == true {
-                if let nextEvent = getNextEvent(events: self.statusBarItem.events) {
-                    scheduleEventNotification(nextEvent)
+            if change.oldValue != change.newValue {
+                NSLog("Changed joinEventNotification from \(change.oldValue) to \(change.newValue)")
+                if change.newValue == true {
+                    if let nextEvent = getNextEvent(events: self.statusBarItem.events) {
+                        scheduleEventNotification(nextEvent)
+                    }
+                } else {
+                    removePendingNotificationRequests()
                 }
-            } else {
-                removePendingNotificationRequests()
             }
         }
         showEventMaxTimeUntilEventThresholdObserver = Defaults.observe(.showEventMaxTimeUntilEventThreshold) { change in
             if change.oldValue != change.newValue {
+                NSLog("Changed showEventMaxTimeUntilEventThreshold from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateTitle()
             }
         }
         showEventMaxTimeUntilEventEnabledObserver = Defaults.observe(.showEventMaxTimeUntilEventEnabled) { change in
             if change.oldValue != change.newValue {
+                NSLog("Change showEventMaxTimeUntilEventEnabled from \(change.oldValue) to \(change.newValue)")
                 self.statusBarItem.updateTitle()
             }
         }